Add an optional order property
Add an order property to each interface in order to create a means of
controlling the ordering of declarations in the interfaces file.  This is
important for sophisticated setups involving bonding, vlans, and bridging where
the order in which interfaces are raised (declaration order) can cause delays
if it's wrong.